+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 ### Changed
 
+- `all-is-cubes-mesh` library:
+    - The return type of `GetBlockMesh::get_block_mesh()` has changed to `Option<&BlockMesh>`.
+      `None` is to be returned when the implementor intends to request that the block be _omitted_ from a produced `SpaceMesh` (such as when it is being rendered separately) rather than the mesh merely being empty (invisible).
+
 ### Removed
 
 ## 0.7.1 (2024-01-27)